+++ src/components/organisms/SIgnUpForm/style.module.css
@@ -1,0 +1,69 @@
+.wrapper {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.container {
+  margin-top: 80px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
+  border: 1px solid #707070;
+  border-radius: 20px;
+  padding: 20px 50px;
+}
+
+.field {
+  display: flex;
+  flex-direction: column;
+  margin: 10px 0;
+}
+
+.label {
+  font-size: 1.5rem;
+  font-weight: 600;
+}
+
+.errorLabel {
+  margin-top: 1px;
+  height: 1.2rem;
+  line-height: 1.2rem;
+  color: red;
+  font-size: 1.2rem;
+  font-weight: 600;
+}
+
+.span {
+  font-size: 1.2rem;
+  font-weight: normal;
+}
+
+.input {
+  font-size: 1.5rem;
+  min-width: 450px;
+  height: 40px;
+  border: 1px solid #707070;
+  border-radius: 5px;
+  padding-left: 10px;
+}
+
+.button {
+  padding: 5px 30px;
+  cursor: pointer;
+  border: none;
+  border-radius: 5px;
+  background-color: #20746a;
+  color: #fff;
+}
+
+.button:disabled {
+  padding: 5px 30px;
+  cursor: not-allowed;
+  border: none;
+  border-radius: 5px;
+  background-color: #20746a;
+  opacity: 0.5;
+  color: #fff;
+}
