+++ src/components/pages/SignUp/style.css
@@ -7,73 +7,3 @@
   font-size: 2.5rem;
   margin-bottom: 0px;
 }
-
-.signup__form--wrapper {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.signup__form--container {
-  margin-top: 80px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-  border: 1px solid #707070;
-  border-radius: 20px;
-  padding: 20px 50px;
-}
-
-.signup__form--field {
-  display: flex;
-  flex-direction: column;
-  margin: 10px 0;
-}
-
-.signup__form--label {
-  font-size: 1.5rem;
-  font-weight: 600;
-}
-
-.signup__form--error-label {
-  margin-top: 1px;
-  height: 1.2rem;
-  line-height: 1.2rem;
-  color: red;
-  font-size: 1.2rem;
-  font-weight: 600;
-}
-
-.signup__form--span {
-  font-size: 1.2rem;
-  font-weight: normal;
-}
-
-.signup__form--input {
-  font-size: 1.5rem;
-  min-width: 450px;
-  height: 40px;
-  border: 1px solid #707070;
-  border-radius: 5px;
-  padding-left: 10px;
-}
-
-.signup__form--button {
-  padding: 5px 30px;
-  cursor: pointer;
-  border: none;
-  border-radius: 5px;
-  background-color: #20746a;
-  color: #fff;
-}
-
-.signup__form--button:disabled {
-  padding: 5px 30px;
-  cursor: not-allowed;
-  border: none;
-  border-radius: 5px;
-  background-color: #20746a;
-  opacity: 0.5;
-  color: #fff;
-}
